Removed display resizing and added bypass buttons
The app is meant to be displayed at 1280x1024, so pages have a max width of 1280, but will be centered properly if the window is larger. For the payment/ticket search pages, there are now buttons on the left to simulate cash deposits and card swipes. Note that you still need to select the payment option before simulating a card swipe.